// src/interfaces/EAction.ts
export enum EAction {
    ADD_UP_ENV = 'ENV/ADD_UP_ENV',
    REMOVE_UP_ENV = 'ENV/REMOVE_UP_ENV',
    ADD_DOWN_ENV = 'ENV/ADD_DOWN_ENV',
    REMOVE_DOWN_ENV = 'ENV/REMOVE_DOWN_ENV',
    ADD_BODY_ITEM = 'ADD_BODY_ITEM',
    ADD_BODY_COSTUME = 'ADD_BODY_COSTUME',
    REMOVE_BODY_ITEM = 'REMOVE_BODY_ITEM',
    ADD_BODY_WEAPON = 'ADD_BODY_WEAPON',
    REMOVE_BODY_WEAPON = 'REMOVE_BODY_WEAPON',
    ADD_INV_ITEM = 'ADD_INV_ITEM',
    REMOVE_INV_ITEM = 'REMOVE_INV_ITEM',
    CHANGE_INV_ITEM_STATE = 'CHANGE_INV_ITEM_STATE',
    CHANGE_BAG_ITEM_STATE = 'CHANGE_BAG_ITEM_STATE',
    ADD_BAG_ITEM = 'ADD_BAG_ITEM',
    REMOVE_BAG_ITEM = 'REMOVE_BAG_ITEM',
    SET_UP_ENV_NAME = 'SET_UP_ENV_NAME',
    SET_DOWN_ENV_NAME = 'SET_DOWN_ENV_NAME',
    REFRESH_UP_ENV_ITEM = 'REFRESH_UP_ENV_ITEM',
    REFRESH_DOWN_ENV_ITEM = 'REFRESH_DOWN_ENV_ITEM',
    REFRESH_BAG_ITEM = 'REFRESH_BAG_ITEM',
    REFRESH_INV_ITEM = 'REFRESH_INV_ITEM',
    ADD_ITEM_DESCRIPTION_INV = 'ADD_ITEM_DESCRIPTION_INV',
    ADD_ITEM_DESCRIPTION_BAG = 'ADD_DESCRIPTION_TO_ENV',
    ADD_DISABLE_ENV_UP = 'ADD_DISABLE_ENV_UP',
    ADD_DISABLE_ENV_DOWN = 'ADD_DISABLE_ENV_DOWN',
    ADD_DISABLE_BODY = 'ADD_DISABLE_BODY',
    ADD_DISABLE_BAG = 'ADD_DISABLE_BAG',
    ADD_DISABLE_INV = 'ADD_DISABLE_INV',
    DISABLE_BAG = 'DISABLE_BAG',
    DISABLE_INV = 'DISABLE_INV',
    SET_BAG_CELL_COUNT = 'SET_BAG_CELL_COUNT',
    SET_INV_CELL_COUNT = 'SET_INV_CELL_COUNT',
    SET_BAG_WEIGHT_LIMIT = 'SET_BAG_WEIGHT_LIMIT',
    SET_INV_WEIGHT_LIMIT = 'SET_INV_WEIGHT_LIMIT',
    SET_INV_CURRENT_WEIGHT = 'SET_INV_CURRENT_WEIGHT',
    SET_BAG_CURRENT_WEIGHT = 'SET_BAG_CURRENT_WEIGHT'
}